+++ atlasdb-dbkvs/src/main/java/com/palantir/atlasdb/keyvalue/dbkvs/OracleTableNameUnmapper.java
@@ -21,9 +21,8 @@ import java.util.concurrent.ExecutionException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
-import com.google.common.cache.CacheLoader;
-import com.google.common.cache.LoadingCache;
 import com.google.common.collect.Iterables;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
@@ -36,48 +35,44 @@ import com.palantir.nexus.db.sql.SqlConnection;
 class OracleTableNameUnmapper {
     private static final Logger log = LoggerFactory.getLogger(OracleTableNameUnmapper.class);
 
-    private final ConnectionSupplier conns;
+    private Cache<String, String> unmappingCache;
 
-    private LoadingCache<String, String> unmappingCache = CacheBuilder.newBuilder().build(
-            new CacheLoader<String, String>() {
-                @Override
-                public String load(String fullTableName) throws Exception {
-                    SqlConnection conn = null;
-                    try {
-                        conn = conns.getNewUnsharedConnection();
-                        AgnosticResultSet results = conn.selectResultSetUnregisteredQuery(
-                                "SELECT short_table_name "
-                                        + "FROM " + AtlasDbConstants.ORACLE_NAME_MAPPING_TABLE
-                                        + " WHERE table_name = ?", fullTableName);
-                        if (results.size() == 0) {
-                            throw new TableMappingNotFoundException(
-                                    "The table " + fullTableName + " does not have a mapping."
-                                            + "This might be because the table does not exist.");
-                        }
+    OracleTableNameUnmapper() {
+        unmappingCache = CacheBuilder.newBuilder().build();
+    }
 
-                        return Iterables.getOnlyElement(results.rows()).getString("short_table_name");
-                    } finally {
-                        if (conn != null) {
-                            try {
-                                conn.getUnderlyingConnection().close();
-                            } catch (SQLException e) {
-                                log.error("Couldn't cleanup SQL connection while performing table name unmapping.", e);
-                            }
+    @SuppressWarnings("checkstyle:NestedTryDepth")
+    public String getShortTableNameFromMappingTable(
+            ConnectionSupplier connectionSupplier,
+            String tablePrefix,
+            TableReference tableRef) throws TableMappingNotFoundException {
+        String fullTableName = tablePrefix + DbKvs.internalTableName(tableRef);
+        try {
+            return unmappingCache.get(fullTableName, () -> {
+                SqlConnection conn = null;
+                try {
+                    conn = connectionSupplier.getNewUnsharedConnection();
+                    AgnosticResultSet results = conn.selectResultSetUnregisteredQuery(
+                            "SELECT short_table_name "
+                                    + "FROM " + AtlasDbConstants.ORACLE_NAME_MAPPING_TABLE
+                                    + " WHERE table_name = ?", fullTableName);
+                    if (results.size() == 0) {
+                        throw new TableMappingNotFoundException(
+                                "The table " + fullTableName + " does not have a mapping."
+                                        + "This might be because the table does not exist.");
+                    }
+
+                    return Iterables.getOnlyElement(results.rows()).getString("short_table_name");
+                } finally {
+                    if (conn != null) {
+                        try {
+                            conn.getUnderlyingConnection().close();
+                        } catch (SQLException e) {
+                            log.error("Couldn't cleanup SQL connection while performing table name unmapping.", e);
                         }
                     }
                 }
-            }
-    );
-
-    OracleTableNameUnmapper(ConnectionSupplier conns) {
-        this.conns = conns;
-    }
-
-    public String getShortTableNameFromMappingTable(String tablePrefix, TableReference tableRef)
-            throws TableMappingNotFoundException {
-        String fullTableName = tablePrefix + DbKvs.internalTableName(tableRef);
-        try {
-            return unmappingCache.get(fullTableName);
+            });
         } catch (ExecutionException e) {
             throw new TableMappingNotFoundException(e.getCause());
         }