+++ docs/source/release_notes/release-notes.rst
@@ -34,6 +34,17 @@ Changelog
 develop
 =======
 
+.. list-table::
+    :widths: 5 40
+    :header-rows: 1
+
+    *    - Type
+         - Change
+
+    *    - |improved|
+         - Oracle query performance improvement; table name cache now global to KVS level.
+         - (`Pull Request <https://github.com/palantir/atlasdb/pull/1325>`__)
+
 .. <<<<------------------------------------------------------------------------------------------------------------->>>>
 
 =======